+++ cmd/workflow-engine/main.go
@@ -1,40 +1,33 @@
 package main
 
 import (
-	"context"
+	"log/slog"
 	"os"
 	"time"
 	"workflow-engine/cmd/workflow-engine/cli"
 
-	"log/slog"
-
-	"github.com/bep/simplecobra"
 	"github.com/lmittmann/tint"
 )
 
 const exitOk = 0
-const exitSystemFailure = 1
-const exitCommandFailure = 2
+const exitUserInput = 1
+const exitSystemFailure = 2
+const exitCommandFailure = 3
+
+const pipelineTypeDebug = "debug"
 
 func main() {
-	leveler := new(slog.LevelVar)
 	// Set up custom structured logging with colorized output
 	slog.SetDefault(slog.New(tint.NewHandler(os.Stderr, &tint.Options{
-		Level:      leveler,
+		Level:      slog.LevelDebug,
 		TimeFormat: time.TimeOnly,
 	})))
 
-	command := cli.NewCommand(leveler)
+	app := cli.NewApp()
 
-	x, err := simplecobra.New(command)
-	if err != nil {
-		os.Exit(exitSystemFailure)
-	}
-
-	_, err = x.Execute(context.Background(), os.Args[1:])
-	if err != nil {
+	if err := app.Execute(); err != nil {
+		slog.Error("command execution failure", "error", err)
 		os.Exit(exitCommandFailure)
 	}
 
-	os.Exit(exitOk)
 }