+++ CHANGELOG.md
@@ -9,8 +9,13 @@ Version header format: `## [x.x.x] - yyyy-mm-dd`
 
 ## [UNRELEASED]
 
+### Changed
+
+- Refactored the directory structure, all pipelines will exist in pkg/pipelines
+
 ### Added
 
+- cmd/workflow-engine for cli
 - pkg/environments
 - pkg/jobs
 - pkg/pipelines